refactor: Bear server implementation and update dependencies
Extract `create_server` to a dedicated module for cleaner separation of concerns. Simplify main file logic, focusing on server initialization and execution. Updated dependencies to ensure compatibility and improve maintainability.
