+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
     hooks:
       - id: mypy
         args: []
-        additional_dependencies: ["mcp>=1.2.1", "pytest>=8.3.4", "fastapi>=0.115.8"]
+        additional_dependencies: ["mcp>=1.3.0", "pytest>=8.3.4", "fastapi>=0.115.8"]
   - repo: local
     hooks:
       - id: pytest
