+++ pyproject.toml
@@ -26,7 +26,7 @@ classifiers = [
 dependencies = [
   "click>=8.1.8",
   "fastapi>=0.115.8",
-  "mcp>=1.3.0",
+  "mcp>=1.3",
   "pydantic>=2.10.6",
   "uvicorn>=0.34",
 ]
